+++  course-work-ps-31-2013 --username dertkaes@gmail.com/OpenGLTest/OpenGLTest/controller/CommandMoveObject.cpp
@@ -2,7 +2,7 @@
 #include "..\model\GameModel.h"
 
 
-CCommandMoveObject::CCommandMoveObject(float deltaX, float deltaY):
+CCommandMoveObject::CCommandMoveObject(double deltaX, double deltaY):
 	m_deltaX(deltaX), m_deltaY(deltaY), m_pObject(CGameModel::GetIntanse().lock()->GetSelectedObject())
 {
 }
