+++  course-work-ps-31-2013 --username dertkaes@gmail.com/OpenGLTest/OpenGLTest/model/MovementLimiter.h
@@ -3,41 +3,41 @@
 class IMoveLimiter
 {
 public:
-	virtual void FixPosition(float & x, float & y, float & z, float & rotation) const = 0;
+	virtual void FixPosition(double & x, double & y, double & z, double & rotation) const = 0;
 	virtual ~IMoveLimiter() {}
 };
 
 class CMoveLimiterCircle : public IMoveLimiter
 {
 public:
-	CMoveLimiterCircle(float x, float y, float radius):m_x(x), m_y(y), m_radius(radius) {}
-	void FixPosition(float & x, float & y, float & z, float & rotation) const;
+	CMoveLimiterCircle(double x, double y, double radius):m_x(x), m_y(y), m_radius(radius) {}
+	void FixPosition(double & x, double & y, double & z, double & rotation) const;
 private:
-	float m_x;
-	float m_y;
-	float m_radius;
+	double m_x;
+	double m_y;
+	double m_radius;
 };
 
 class CMoveLimiterRectangle : public IMoveLimiter
 {
 public:
-	CMoveLimiterRectangle(float x1, float y1, float x2, float y2):m_minX((x1 < x2)?x1:x2), m_maxX((x1 > x2)?x1:x2), m_minY((y1 < y2)?y1:y2), m_maxY((y1 > y2)?y1:y2) {}
-	void FixPosition(float & x, float & y, float & z, float & rotation) const;
+	CMoveLimiterRectangle(double x1, double y1, double x2, double y2):m_minX((x1 < x2)?x1:x2), m_maxX((x1 > x2)?x1:x2), m_minY((y1 < y2)?y1:y2), m_maxY((y1 > y2)?y1:y2) {}
+	void FixPosition(double & x, double & y, double & z, double & rotation) const;
 private:
-	float m_minX;
-	float m_minY;
-	float m_maxX;
-	float m_maxY;
+	double m_minX;
+	double m_minY;
+	double m_maxX;
+	double m_maxY;
 };
 
 class CMoveLimiterStatic : public IMoveLimiter
 {
 public:
-	CMoveLimiterStatic(float x, float y, float z, float rotation):m_x(x), m_y(y), m_z(z), m_rotation(rotation) {}
-	void FixPosition(float & x, float & y, float & z, float & rotation) const;
+	CMoveLimiterStatic(double x, double y, double z, double rotation):m_x(x), m_y(y), m_z(z), m_rotation(rotation) {}
+	void FixPosition(double & x, double & y, double & z, double & rotation) const;
 private:
-	float m_x;
-	float m_y;
-	float m_z;
-	float m_rotation;
+	double m_x;
+	double m_y;
+	double m_z;
+	double m_rotation;
 };
