+++  course-work-ps-31-2013 --username dertkaes@gmail.com/OpenGLTest/OpenGLTest/view/SkyBox.h
@@ -3,11 +3,11 @@
 class CSkyBox
 {
 public:
-	CSkyBox(float width, float height, float length, std::string const& imageFolder);
-	void Draw(float x, float y, float z, float scale);
+	CSkyBox(double width, double height, double length, std::string const& imageFolder);
+	void Draw(double x, double y, double z, double scale);
 private:
-	float m_height;
-	float m_width;
-	float m_length;
+	double m_height;
+	double m_width;
+	double m_length;
 	std::string m_imageFolder;
 };
